Move all CreateBatchAllocations filtes to its folder
